move cursor to its own file
